Adjust some files
