+++ p2pserver/net/transport/quic/transport.go
@@ -24,14 +24,12 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/pem"
-	"io"
 	"math/big"
 	"net"
 	"time"
 
 	"github.com/lucas-clemente/quic-go"
 	"github.com/ontio/ontology/common/config"
-	"github.com/ontio/ontology/common/log"
 	"github.com/ontio/ontology/p2pserver/common"
 	tsp "github.com/ontio/ontology/p2pserver/net/transport"
 )
@@ -77,65 +75,8 @@ var quicConfig = &quic.Config{
 	KeepAlive: false,
 }
 
-type connection struct {
-	sess          quic.Session
-	streamWTimeOut time.Time
-}
-
 type transport struct {
 	tlsConf *tls.Config
-}
-
-func (this * connection) GetReader() (io.Reader, error) {
-
-	stream, err := this.sess.AcceptUniStream()
-	if err != nil {
-		log.Errorf("[p2p]AcceptStream lAddr=%s, rAddr=%s, ERR:%s", this.sess.LocalAddr().String(), this.sess.RemoteAddr().String(), err)
-		return  nil, err
-	}
-
-	return stream, nil
-}
-
-func (this * connection) Write(b []byte) (int, error) {
-
-	stream, err := this.sess.OpenUniStreamSync()
-	if err != nil {
-		log.Errorf("[p2p]OpenStreamSync lAddr=%s, rAddr=%s, ERR:%s", this.sess.LocalAddr().String(), this.sess.RemoteAddr().String(), err)
-		return 0, err
-	}
-	defer stream.Close()
-
-	stream.SetWriteDeadline(this.streamWTimeOut)
-	cntW, errW := stream.Write(b)
-	if errW != nil {
-		log.Errorf("[p2p] Write err by stream:%s", errW)
-		return 0, errW
-	}
-
-	return cntW, errW
-}
-
-func (this * connection) Close() error {
-
-	return this.sess.Close()
-}
-
-func (this* connection) LocalAddr() net.Addr {
-
-	return this.sess.LocalAddr()
-}
-
-func (this* connection) RemoteAddr() net.Addr {
-
-	return this.sess.RemoteAddr()
-}
-
-func (this * connection) SetWriteDeadline(t time.Time) error {
-
-	this.streamWTimeOut = t
-
-	return nil
 }
 
 func NewTransport( ) (tsp.Transport, error) {
