+++ p2pserver/net/transport/tcp/connection.go
@@ -15,26 +15,45 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with The ontology.  If not, see <http://www.gnu.org/licenses/>.
  */
-package creator
+
+package tcp
 
 import (
-	"github.com/ontio/ontology/common/log"
-	"github.com/ontio/ontology/p2pserver/common"
-	"testing"
+	"io"
+	"net"
+	"time"
 )
 
-func init() {
-	log.Init(log.Stdout)
+type connection struct {
+	net.Conn
+	io.Reader
 }
 
-func TestFactory(t *testing.T) {
-	tcpTransport1, _ := GetTransportFactory().GetTransport(10)
-	if tcpTransport1 != nil {
-		t.Error("tcpTransport1 should be nil")
-	}
+func (this * connection) GetReader() (io.Reader, error) {
 
-	tcpTransport2, _ := GetTransportFactory().GetTransport(common.T_TCP)
-	if tcpTransport2 == nil {
-		t.Error("tcpTransport2 shouldnot be nil")
-	}
+	return  this.Reader, nil
+}
+
+func (this * connection) Write(b []byte) (n int, err error) {
+
+	return this.Conn.Write(b)
+}
+
+func (this * connection) Close() error {
+
+	return  this.Conn.Close()
+}
+
+func (this* connection) LocalAddr() net.Addr {
+
+	return this.Conn.LocalAddr()
+}
+
+func (this* connection) RemoteAddr() net.Addr {
+
+	return this.Conn.RemoteAddr()
+}
+func (this * connection) SetWriteDeadline(t time.Time) error {
+
+	return  this.Conn.SetWriteDeadline(t)
 }
